requirements.txt
Flask==2.1.2
web3==5.28.0
ipfshttpclient==0.8.0a2
python-dotenv==0.20.0
